# drone-backend/src/main/java/com/uav/admin/controller/AdminController.java
package com.uav.admin.controller;

import com.uav.admin.pojo.dto.AdminDto;
import com.uav.admin.pojo.entity.Admin;
import com.uav.server.result.Result;
import com.uav.admin.pojo.vo.AdminLoginVO;
import com.uav.server.annotation.OperationLog;
import com.uav.server.annotation.SkipJwt;
import com.uav.server.util.JwtUtil;
import com.uav.server.util.UserContext;
import com.uav.admin.service.AdminService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

@Tag(name = "Admin Auth API")
@RestController
@RequestMapping("/admin")
public class AdminController {

    @Autowired
    private AdminService adminService;

    @Autowired
    private JwtUtil jwtUtil;

    @SkipJwt
    @OperationLog("管理员登录")
    @Operation(
            summary = "管理员登录",
            description = "管理员登录接口",
            responses = {
                    @ApiResponse(responseCode = "200", description = "登录成功"),
                    @ApiResponse(responseCode = "401", description = "用户名或密码错误")
            }
    )
    @PostMapping("/login")
    public Result<AdminLoginVO> adminLogin(@RequestBody AdminDto adminDto) {
        Admin admin = adminService.tryToLogin(adminDto);
        UserContext.setUser(admin.getId(), admin.getName(), 2);

        AdminLoginVO.AdminInfo adminInfo = new AdminLoginVO.AdminInfo();
        adminInfo.setId(admin.getId());
        adminInfo.setName(admin.getName());
        adminInfo.setPhoneNumber(admin.getPhoneNumber());

        AdminLoginVO vo = new AdminLoginVO();
        vo.setToken(jwtUtil.generateToken(admin.getId(), admin.getName(), 2));
        vo.setAdmin(adminInfo);

        return Result.success("登录成功", vo);
    }
}