# spec/openapi/drone-backend.openapi.json
{
  "components" : {
    "schemas" : {
...
      "AdminDto" : {
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "AdminLoginVO" : {
        "properties" : {
          "admin" : {
            "$ref" : "#/components/schemas/AdminInfo"
          },
          "token" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "OrderListVO" : {
        "properties" : {
          "currentPage" : {
            "format" : "int32",
            "type" : "integer"
          },
          "orders" : {
            "items" : {
              "$ref" : "#/components/schemas/OrderVO"
            },
            "type" : "array"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ResultAdminLoginVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdminLoginVO",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ResultAdminOrderVo" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdminOrderVo",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ResultAdminPageVoAdminOrderVo" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdminPageVoAdminOrderVo",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ResultAdminPageVoAdminTaskVo" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdminPageVoAdminTaskVo",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultAdminTaskVo" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdminTaskVo",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultListMapStringObject" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）",
            "items" : {
              "additionalProperties" : { },
              "type" : "object"
            },
            "type" : "array"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultMapStringObject" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "additionalProperties" : { },
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）",
            "type" : "object"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ResultMapStringString" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）",
            "type" : "object"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultOrderListVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/OrderListVO",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultOrderVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/OrderVO",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultTaskPageVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/TaskPageVO",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ResultTaskVo" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/TaskVo",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "Bearer Authentication" : {
        "bearerFormat" : "JWT",
        "description" : "请输入JWT Token，格式：直接输入token字符串即可",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "无人机项目web后端接口文档",
    "title" : "无人机项目接口文档",
    "version" : "1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/admin/login" : {
      "post" : {
        "description" : "管理员登录接口",
        "operationId" : "adminLogin",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AdminDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultAdminLoginVO"
                }
              }
            },
            "description" : "登录成功"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultAdminLoginVO"
                }
              }
            },
            "description" : "用户名或密码错误"
          }
        },
        "summary" : "管理员登录",
        "tags" : [ "Admin Auth API" ]
      }
    },
...
    "/admin/orders" : {
      "get" : {
        "description" : "支持按订单状态（枚举名或状态码）与订单号/任务号精确过滤；page 从 0 起",
        "operationId" : "listOrders_1",
        "parameters" : [ {
          "description" : "任务编号精确过滤",
          "in" : "query",
          "name" : "taskNum",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "每页条数（默认 20，上限 100）",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "description" : "订单号精确过滤",
          "in" : "query",
          "name" : "orderNum",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "订单状态过滤：枚举名（PAID）或状态码（1）",
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "页码（从 0 起）",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultAdminPageVoAdminOrderVo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "全平台订单分页列表",
        "tags" : [ "Admin Query API" ]
      }
    },
    "/admin/orders/{orderNum}" : {
      "get" : {
        "operationId" : "orderDetail",
        "parameters" : [ {
          "description" : "订单号",
          "in" : "path",
          "name" : "orderNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultAdminOrderVo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "订单详情",
        "tags" : [ "Admin Query API" ]
      }
    },
    "/admin/tasks" : {
      "get" : {
        "description" : "支持按任务状态（枚举名：IDLE/IN_PROGRESS/COMPLETED）与任务号精确过滤；含订单双状态与操作提示",
        "operationId" : "listTasks_1",
        "parameters" : [ {
          "description" : "任务状态过滤：IDLE / IN_PROGRESS / COMPLETED",
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "任务编号精确过滤",
          "in" : "query",
          "name" : "taskNum",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "每页条数（默认 20，上限 100）",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "description" : "页码（从 0 起）",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultAdminPageVoAdminTaskVo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "全平台任务分页列表",
        "tags" : [ "Admin Query API" ]
      }
    },
    "/admin/tasks/{taskNum}" : {
      "get" : {
        "description" : "含订单双状态、飞手与完成说明、操作提示",
        "operationId" : "taskDetail",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "path",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultAdminTaskVo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "任务详情",
        "tags" : [ "Admin Query API" ]
      }
    },
...
    "/order/detail" : {
      "get" : {
        "description" : "根据订单号获取详细信息（含交付文件列表）",
        "operationId" : "getOrderDetail",
        "parameters" : [ {
          "description" : "订单号",
          "in" : "query",
          "name" : "orderNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultOrderVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "订单详情",
        "tags" : [ "Order API" ]
      }
    },
...
    "/order/list" : {
      "get" : {
        "description" : "获取当前用户的所有订单，按创建时间倒序",
        "operationId" : "listOrders",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultOrderListVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "订单列表",
        "tags" : [ "Order API" ]
      }
    },
...
    "/rider/task/detail" : {
      "get" : {
        "description" : "飞手查看任意任务的详细信息（无需任务归属）",
        "operationId" : "getTaskDetail_1",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "query",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTaskVo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "任务详情",
        "tags" : [ "Rider API" ]
      }
    },
...
    "/task/detail" : {
      "get" : {
        "description" : "根据任务编号获取详细信息，包含航点列表",
        "operationId" : "getTaskDetail",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "query",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTaskVo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "获取任务详情",
        "tags" : [ "Task API" ]
      }
    },
...
    "/task/list" : {
      "get" : {
        "operationId" : "listTasks",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTaskPageVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "获取当前用户的任务列表",
        "tags" : [ "Task API" ]
      }
    },
...
    "/tasks/{taskNum}/attachments" : {
      "get" : {
        "description" : "任务所有者与接单飞手可看；逐条附 presigned 下载 URL（15 分钟有效）",
        "operationId" : "list_1",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "path",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListMapStringObject"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "附件列表",
        "tags" : [ "Task Attachment API" ]
      }
    },
    "/tasks/{taskNum}/attachments/download-url" : {
      "get" : {
        "description" : "任务所有者或接单飞手可看",
        "operationId" : "downloadUrl",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "path",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "附件对象键",
          "in" : "query",
          "name" : "objectKey",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultMapStringObject"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "获取交付物下载凭证",
        "tags" : [ "Task Attachment API" ]
      }
    },
    "/tasks/{taskNum}/attachments/upload-url" : {
      "post" : {
        "description" : "仅任务的接单飞手可调用；校验类型（图片 jpeg/png/webp ≤20MB，视频 mp4/mov ≤500MB）后返回 presigned PUT URL（15 分钟有效），客户端直传 MinIO",
        "operationId" : "uploadUrl",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "path",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "内容类型（image/jpeg 等）",
          "in" : "query",
          "name" : "contentType",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "原始文件名",
          "in" : "query",
          "name" : "fileName",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "文件大小（字节）",
          "in" : "query",
          "name" : "sizeBytes",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultMapStringObject"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "获取交付物上传凭证",
        "tags" : [ "Task Attachment API" ]
      }
    },
...
    "/upload/download/{id}" : {
      "get" : {
        "description" : "返回 OSS 预签名 GET URL，30 分钟有效",
        "operationId" : "download",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultMapStringString"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "获取文件下载链接",
        "tags" : [ "Upload API" ]
      }
    },
...
  },
  "security" : [ {
    "Bearer Authentication" : [ ]
  } ],
  "servers" : [ {
    "description" : "相对基址：由部署环境（网关/反向代理）决定",
    "url" : "/"
  } ],
...
}